in progress of updating part of the dependencies and using cmake fetch

+++ Engine/src/Buffers/CommandBuffers/CommandPool.h
@@ -18,7 +18,7 @@ namespace Rapture {
         uint32_t queueFamilyIndex;
         VkCommandPoolCreateFlags flags=VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT;
 
-        uint32_t hash() const {
+        size_t hash() const {
             return std::hash<size_t>()(threadId) ^ std::hash<uint32_t>()(queueFamilyIndex) ^ std::hash<VkCommandPoolCreateFlags>()(flags);
         }
     };
